requirements.txt
##Python 3.10.6
streamlit==1.14.0
yt-dlp==2022.10.4
ffmpeg-python==0.2.0